requirements.txt
fastapi
boto3
Pillow
selenium
mangum==0.10.0b1